+++ CHANGELOG.md
@@ -11,6 +11,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Replace external redmine-java-api with internal implementation ([#15](https://github.com/scm-manager/scm-redmine-plugin/pull/15))
 
+### Fixed
+- Ignore repository-level configuration if it is disabled in the global configuration ([#14](https://github.com/scm-manager/scm-redmine-plugin/pull/14))
+
 ## [2.0.0] - 2020-06-04
 ### Changed
 - Changeover to MIT license ([#4](https://github.com/scm-manager/scm-redmine-plugin/pull/4))